refactor: rename project name
