+++ src/components/BookshelfLogo/BookshelfLogo.tsx
@@ -1,13 +1,13 @@
 import { Title } from '@mantine/core'
 import Link from 'next/link'
-import { useStyles } from './BookshelfieLogo.styles'
+import { useStyles } from './BookshelfLogo.styles'
 
-export function BookshelfieLogo() {
+export function BookshelfLogo() {
   const { classes } = useStyles()
 
   return (
     <Link href="/" className={classes.link}>
-      <Title order={1}>📓 bookshelfie</Title>
+      <Title order={1}>📓 bookshelf</Title>
     </Link>
   )
 }
