+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export * from './BookshelfieLogo/BookshelfieLogo'
+export * from './BookshelfLogo/BookshelfLogo'
 export * from './Header/Header'
 export * from './Page/Page'
